+++ src/main/kotlin/app/CreateUser.kt
@@ -1,4 +1,4 @@
-package features
+package app
 
 import domain.User
 import domain.UserRepository
