+++ src/main/kotlin/app/ListUsers.kt
@@ -1,4 +1,4 @@
-package features
+package app
 
 import domain.UserRepository
