+++ src/test/kotlin/adapters/web/IntegrationTest.kt
@@ -1,4 +1,4 @@
-package web
+package adapters.web
 
 import domain.IdGenerator
 import io.mockk.every
