+++ src/test/kotlin/adapters/web/IntegrationTestWithMongoDB.kt
@@ -1,35 +1,37 @@
-package web
+package adapters.web
 
-import com.wix.mysql.EmbeddedMysql
-import com.wix.mysql.EmbeddedMysql.anEmbeddedMysql
-import com.wix.mysql.config.MysqldConfig.aMysqldConfig
-import com.wix.mysql.distribution.Version
+import de.flapdoodle.embed.mongo.MongodExecutable
+import de.flapdoodle.embed.mongo.MongodProcess
+import de.flapdoodle.embed.mongo.MongodStarter
+import de.flapdoodle.embed.mongo.config.MongodConfigBuilder
+import de.flapdoodle.embed.mongo.config.Net
+import de.flapdoodle.embed.mongo.distribution.Version
+import de.flapdoodle.embed.process.runtime.Network
 import domain.UserRepository
-import org.jetbrains.exposed.sql.Database
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import persistence.MySqlUserRepository
+import persistence.MongoDBUserRepository
 
-object IntegrationTestWithMySql {
+object IntegrationTestWithMongoDB {
 
     private lateinit var webAppConfig: WebAppConfig
-    private lateinit var dbServer: EmbeddedMysql
+    private lateinit var mongodExe: MongodExecutable
+    private lateinit var mongod: MongodProcess
     private lateinit var userRepository: UserRepository
 
     @BeforeAll
     @JvmStatic
     fun setup() {
-        val config = aMysqldConfig(Version.v5_7_latest).withPort(3301).withUser("user", "pass").build()
-        dbServer = anEmbeddedMysql(config).addSchema("test_schema").start()
-        userRepository = MySqlUserRepository(
-            Database.connect(
-                url = "jdbc:mysql://user:pass@localhost:3301/test_schema",
-                driver = "com.mysql.cj.jdbc.Driver"
-            )
+        mongodExe = MongodStarter.getDefaultInstance().prepare(
+            MongodConfigBuilder()
+                .version(Version.Main.PRODUCTION)
+                .net(Net("localhost", 12345, Network.localhostIsIPv6()))
+                .build()
         )
-        userRepository.createSchema()
+        mongod = mongodExe.start()
+        userRepository = MongoDBUserRepository("localhost", 12345, "db123").apply { createSchema() }
         webAppConfig = WebAppConfig(userRepository, 8081).apply { start() }
     }
 
@@ -52,6 +54,7 @@ object IntegrationTestWithMySql {
     @JvmStatic
     fun afterAll() {
         webAppConfig.stop()
-        dbServer.stop()
+        mongod.stop()
+        mongodExe.stop()
     }
 }
